chore: move platform-specific types and enums to platform folder
